Fix issue with fetch using previous data for first request after change
Basically, when a user changes the currency, the first fetch that fires after that still uses the currencies that were selected just prior to the change. The second fetch request after the currency changes will use the newly selected data.
The issue is due to the function reference being passed to `requestAnimationFrame` not being updated until a fetch has run.
This change forces a new requestAnimationFrame to be registered because the `useEffect` will be triggered when the `fetchData` function changes (i.e. when it gets new to/from currency values)